Spring cleaning (6): burnout is real.
- Removed initialization of `RenderingCore` from `DisplayManager`.
- Use `std::(u)intXX` for fixed-width integers instead of rolling out our own.
- Fixed issue with `Ssize` being unsigned.
- Removed `WindowHandle` and `ScreenHandle` type aliases.
- Added support for using range-based for loops with `Set`.
- Fixed `FileStream` always flushing on destruction, even when the
  stream is read-only.
- Got the codebase compiling in MSVC.

+++ Source/Runtime/Foundation/Windows/WindowsConsoleWriter.cpp
@@ -17,7 +17,7 @@ namespace Kitsune::Details
 
         WideString wideString = Utf8ToUtf16<char, wchar_t>(string);
         ::WriteConsoleW(
-            handle, wideString.Data(), wideString.Size(),
+            handle, wideString.Data(), static_cast<DWORD>(wideString.Size()),
             nullptr, nullptr);
     }
 }
